Move 'ConstantPool' to program.h/c.
