+++ src/program.h
@@ -31,6 +31,16 @@ typedef enum {
 } Opcode;
 
 typedef struct {
+  ThuslyValue* values;
+  int count;
+  int capacity;
+} ConstantPool;
+
+void constant_pool_init(ConstantPool* pool);
+void constant_pool_free(ConstantPool* pool);
+void constant_pool_add(ConstantPool* pool, ThuslyValue value);
+
+typedef struct {
   ConstantPool constant_pool;
   int* source_lines; // Mirrors the `instructions`
   byte* instructions;
